Refactor: FileUtil생성 및 User쪽 리팩토링
