+++ src/main/java/com/sprint/mission/discodeit/controller/dto/UserReadStatusResponse.java
@@ -3,7 +3,9 @@ package com.sprint.mission.discodeit.controller.dto;
 import com.sprint.mission.discodeit.entity.ReadStatus;
 import java.time.Instant;
 import java.util.UUID;
+import lombok.Builder;
 
+@Builder
 public record UserReadStatusResponse(
     UUID id,
     Instant createdAt,
@@ -14,8 +16,13 @@ public record UserReadStatusResponse(
 ) {
 
   public static UserReadStatusResponse of(ReadStatus readStatus) {
-    return new UserReadStatusResponse(readStatus.getId(),
-        readStatus.getCreatedAt(), readStatus.getUpdatedAt(),
-        readStatus.getUserId(), readStatus.getChannelId(), readStatus.getLastReadAt());
+    return UserReadStatusResponse.builder()
+        .id(readStatus.getId())
+        .createdAt(readStatus.getCreatedAt())
+        .updatedAt(readStatus.getUpdatedAt())
+        .userId(readStatus.getUserId())
+        .channelId(readStatus.getChannelId())
+        .lastReadAt(readStatus.getLastReadAt())
+        .build();
   }
 }
